refactor: userAgentのschema名修正、schemaの型をfeature/types内に定義
