+++ frontend/src/features/userAgent/index.ts
@@ -1,4 +1,4 @@
-export type { UserAgentFormRequest } from './components/UserAgentForm'
 export * from './components/CreateUserAgentButton'
 export * from './components/UpdateUserAgentButton'
 export * from './useUserAgent'
+export * from './types'
