+++ frontend/src/features/userAgent/types.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod'
+
+import { userAgentRequestSchema } from '@/libs/validates'
+
+export type UserAgentFormRequest = z.infer<typeof userAgentRequestSchema>
