+++ frontend/src/pages/api/useragents/[id].ts
@@ -1,7 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
 
 import prisma from '@/libs/prisma'
-import { apiValidation, userAgentSchema } from '@/libs/validates'
+import { apiValidation, userAgentRequestSchema } from '@/libs/validates'
 
 export default async function handler(
   req: NextApiRequest,
@@ -26,7 +26,7 @@ export default async function handler(
       break
 
     case 'PUT':
-      apiValidation(req, res, userAgentSchema, async () => {
+      apiValidation(req, res, userAgentRequestSchema, async () => {
         const updatedUserAgent = await prisma.userAgent.update({
           where: {
             id: id,
